Improved performance for gallery on desktops with very large files.

+++ Ports/JavaSE/src/com/codename1/impl/javase/JavaSEPort.java
@@ -5974,16 +5974,20 @@ public class JavaSEPort extends CodenameOneImplementation {
                 File selected = pickFile(imageTypes, desc);
 
                 com.codename1.ui.events.ActionEvent result = null;
-                if (selected != null) {
-                    try {
-                        File tmp = File.createTempFile("temp", "." + imageTypes[0]);
-                        tmp.deleteOnExit();
-                        copyFile(selected, tmp);
-                        result = new com.codename1.ui.events.ActionEvent("file:/" + tmp.getAbsolutePath().replace('\\', '/'));
-                    } catch(IOException err) {
-                        err.printStackTrace();
-                    }
-                } 
+                if(exposeFilesystem) { 
+                    if (selected != null) {
+                        try {
+                            File tmp = File.createTempFile("temp", "." + imageTypes[0]);
+                            tmp.deleteOnExit();
+                            copyFile(selected, tmp);
+                            result = new com.codename1.ui.events.ActionEvent("file:/" + tmp.getAbsolutePath().replace('\\', '/'));
+                        } catch(IOException err) {
+                            err.printStackTrace();
+                        }
+                    } 
+                } else {
+                    result = new com.codename1.ui.events.ActionEvent("file:/" + selected.getAbsolutePath().replace('\\', '/'));
+                }
                 final com.codename1.ui.events.ActionEvent finalResult = result;
                 Display.getInstance().callSerially(new Runnable() {
